0.1.0 release

+++ src/main/java/net/xXinailXx/thirteen_flames/item/base/tools/DiggerItemTF.java
@@ -28,8 +28,8 @@ public abstract class DiggerItemTF extends TieredItemTF implements Vanishable {
         this.speed = tier.getSpeed();
         this.attackDamageBaseline = attackDamage + tier.getAttackDamageBonus();
         ImmutableMultimap.Builder<Attribute, AttributeModifier> builder = ImmutableMultimap.builder();
-        builder.put(Attributes.ATTACK_DAMAGE, new AttributeModifier(BASE_ATTACK_DAMAGE_UUID, "Tool modifier", (double)this.attackDamageBaseline, AttributeModifier.Operation.ADDITION));
-        builder.put(Attributes.ATTACK_SPEED, new AttributeModifier(BASE_ATTACK_SPEED_UUID, "Tool modifier", (double)speedMining, AttributeModifier.Operation.ADDITION));
+        builder.put(Attributes.ATTACK_DAMAGE, new AttributeModifier(BASE_ATTACK_DAMAGE_UUID, "Tool modifier", this.attackDamageBaseline, AttributeModifier.Operation.ADDITION));
+        builder.put(Attributes.ATTACK_SPEED, new AttributeModifier(BASE_ATTACK_SPEED_UUID, "Tool modifier", speedMining, AttributeModifier.Operation.ADDITION));
         this.defaultModifiers = builder.build();
     }
